pyproject.toml
[tool.poetry]
name = "mechvibes-lite"
version = "2.1.0"
description = "Lightweight CLI version of Mechvibes by hainguyents13 using Python."
authors = ["eeriemyxi <eeriemyxi@proton.me>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "mechvibes" }]

[tool.poetry.dependencies]
python = "^3.10"
pyglet = "^2.0.9"
rich = "^13.6.0"
click = "^8.1.7"
pyyaml = "^6.0.1"
mergedeep = "^1.3.4"
keyboard = { version = "^0.13.5", markers = "sys_platform == 'win32'" }
evdev = { version = "^1.6.1", markers = "sys_platform == 'linux'" }

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "^23.9.1"
ruff = "^0.0.292"
pyright = "^1.1.331"
typing-extensions = "^4.8.0"

[tool.poetry.scripts]
mvibes = "mechvibes.cli.core:main"

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"

[tool.ruff]
# select = ["E4", "E7", "E9", "F", "I"]
ignore = [
    "ANN1", "ANN204", "ANN401",
    "A003",
    "D1", "D203", "D213"
]
select = ["A", "ANN", "B", "C4", "C90", "D", "E", "F", "I", "ISC", "INT", "N", "PIE", "PT", "Q", "RET", "RUF", "S", "SIM", "TID", "T20", "UP", "W"]

[tool.ruff.per-file-ignores]
# Allow assert in tests
"tests/*" = ["ANN", "S101"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
